requirements.txt
openai
python-dotenv
pdf2image
